split club/tests.py
